Moved helper fns to a new directory. Added test for stars
